test: update label-shape test case
