+++ packages/g6/__tests__/demo/static/index.ts
@@ -1,2 +1,2 @@
-export * from './label';
+export * from './label-shape';
 export * from './layered-canvas';
